eliminate compiler warning

+++ coders/thumbnail.c
@@ -200,7 +200,7 @@ static MagickBooleanType WriteTHUMBNAILImage(const ImageInfo *image_info,
     q++;
   }
   if ((q > (GetStringInfoDatum(profile)+GetStringInfoLength(profile))) ||
-      (length > (GetStringInfoDatum(profile)+GetStringInfoLength(profile)-q)))
+      ((ssize_t) length > (GetStringInfoDatum(profile)+GetStringInfoLength(profile)-q)))
     ThrowWriterException(CoderError,"ImageDoesNotHaveAThumbnail");
   thumbnail_image=BlobToImage(image_info,q,length,exception);
   if (thumbnail_image == (Image *) NULL)
